AdminUser Migration deleted, not longer needed

+++ db/migrate/20120915141850_delete_admin_user.rb
@@ -1,6 +1,6 @@
 class DeleteAdminUser < ActiveRecord::Migration
   def up
-    drop_table :admin_users
+    drop_table :admin_users if ActiveRecord::Base.connection.tables.include?("admin_users")
   end
 
   def down
